Refactor: 필터 초기화 모듈화
