+++ Week 02/🔥assignment/index.html
@@ -133,6 +133,7 @@
       </form>
     </dialog>
     <script type="module" src="./js/index.js"></script>
+    <script type="module" src="./js/modules/filter.js"></script>
     <script type="module" src="./js/modules/addData.js"></script>
   </body>
 </html>
